# References and relevant issues
I work with large tracking datasets (~10⁵ tracks, ~10⁶ vertices), and creating/setting data of a Tracks layer is rather slow, on the order of 10 seconds on my machine). I have dug into the code and found that the culprit is this function in `napari/layers/tracks/_track_utils.py`

```python
def build_tracks(self):
        """build the tracks"""

        points_id = []
        track_vertices = []
        track_connex = []

        # NOTE(arl): this takes some time when the number of tracks is large
        for idx in self.unique_track_ids:
            indices = self._vertex_indices_from_id(idx)

            # grab the correct vertices and sort by time
            vertices = self.data[indices, 1:]

            # coordinates of the text identifiers, vertices and connections
            points_id += [idx] * vertices.shape[0]
            track_vertices.append(vertices)
            track_connex.append(connex(vertices))

        self._points_id = np.array(points_id)[self._ordered_points_idx]
        self._track_vertices = np.concatenate(track_vertices, axis=0)
        self._track_connex = np.concatenate(track_connex, axis=0)
```
Note the comment above the for-loop (which I believe comes from @quantumjot ). The line `indices = self._vertex_indices_from_id(idx)`, which finds vertices indices in a lookup table (a sparse matrix), slows the loop quite a bit. I believe the loop can be circumvented entirely.

# Description
`build_tracks` is called every time a Tracks layer's data is set (so when it is created, or updated after creation). The particular point of the function is to set  `self._points_id`, `self._track_vertices`, and `self._track_connex`,  which are later fed to Vispy to display the tracks. After inspecting these quantities, I think they could be computed faster (without a for-loop).
 *  `self._points_id` is a numpy array representing the track id associated to all vertices sorted by time.
 *  `self._track_vertices` is a numpy array representing the vertices coordinates
 * `self._track_connex` is a numpy array whose length is equal to the number of vertices, and which is filled with `True` except for vertices that end a track, for which it is equal to `False`

### 1. `self._points_id`
In the current version, a list `points_id` is constructed by looping over all unique tracks indices `idx` and incrementing the list by a sub_list `[idx] * vertices.shape[0]`. The final list is then converted as a numpy array and sorted with respect to time using `np.array(points_id)[self._ordered_points_idx]`.
This can be replaced by directly defining `points_id` as `points_id = self.data[:, 0][self._ordered_points_idx].copy()`, i.e directly requesting the tracks indices from the original Tracks layer data.

### 2. self._track_vertices
Like in the previous case, a variable `vertices = self.data[indices, 1:]` is defined during the loop and appended to a list, and the latter is concatenated into `self._track_vertices`. Here, `indices` are the vertex indices, obtained by querying a lookup table. The lookup table is built in the `data.setter` of the same script as
```python
self._id2idxs = coo_matrix(
    (
        np.broadcast_to(1, self.track_ids.size),  # just dummy ones
        (self.track_ids, np.arange(self.track_ids.size)),
    )
).tocsr() 
```
Note that in the `data.setter`, the Tracks layers data `self.data` is first conveniently sorted by ID then time. This means that
`self.track_ids` (which is `self.data[:, 0].astype(np.uint32)`) is sorted. The current implementation can thus be replaced by `track_vertices = self.data[:, 1:].copy()`, i.e the vertices coordinates of the Tracks layer's original data. 

### 3. self._track_connex
Again, a list `track_connex` is created and is appended during the loop (`track_connex.append(connex(vertices))`), then concatenated to define `self._track_connex.` The function `connex` is defined as
```python
def connex(vertices: np.ndarray) -> list:
    """Connection array to build vertex edges for vispy LineVisual.

    Notes
    -----
    See
    http://api.vispy.org/en/latest/visuals.html#vispy.visuals.LineVisual

    """
    return [True] * (vertices.shape[0] - 1) + [False] 
```
Note that it uses its argument `vertices` only to compute `vertices.shape[0]`, i.e the track's duration in frames (the number of vertices in the track). `track_connex` can be defined directly by
```python
# Indices in the data array just before the track id changes
indices_new_id = np.where(np.diff(self.data[:, 0]))[0]

# Define track_connex as an array full of 'True', then set to 'False'
# at the indices just before the track id changes
track_connex = np.ones(self.data.shape[0], dtype=bool)
track_connex[indices_new_id] = False
```

# Performance changes
I have tested the performance change on my machine using the following script, which simulates a random tracking dataset with `N_vertices` vertices and `N_tracks` tracks.
```python
import numpy as np
import napari
from time import time

np.random.seed(2024)

N_vertices = 1_000_000
N_tracks = 100_000

# simulate 3D tracks (tracks are immediately sorted by track id
# and time in view_tracks/add_tracks, so we can define them randomly)
tracks = np.random.rand(N_vertices, 5) * 100
# make integer track ids
tracks[:,0] = np.random.randint(0, N_tracks, N_vertices)

tracks2 = np.random.rand(N_vertices, 5) * 100
tracks2[:,0] = np.random.randint(0, N_tracks, N_vertices)

viewer = napari.Viewer()
t0 = time()
viewer.add_tracks(tracks, name='tracks')
print(f"Creating tracks with {N_vertices} vertices, {N_tracks} tracks: {time() - t0:.4f} seconds")

t0 = time()
viewer.layers['tracks'].data = tracks2
print(f"Setting tracks with {N_vertices} vertices, {N_tracks} tracks: {time() - t0:.4f} seconds")
```
For different values of `N_vertices` and `N_tracks`, I time the following duration to define and update the Tracks layer:
|  | Old implementation  | New implementation |
| ------------- | ------------- | ------------- |
| `N_vertices = 1_000`, `N_tracks = 100`  | creation:**0.22s**, update:**0.03s**   | creation:**0.17**s, update:**0.008**s  |
| `N_vertices = 1_000_000`, `N_tracks = 10_000`  | creation:**3.49s**, update:**1.86s**   | creation:**2.81**s, update:**0.99**s  |
| `N_vertices = 1_000_000`, `N_tracks = 100_000`  | creation:**10.53s**, update:**8.94s**  | creation:**2.83s**, update:**1.03s**  |

Note that the proposed implementation has a much nicer scaling with the number of tracks and vertices, even for small values of both.
I don't know how to make this benchmark into a proper test though.

# Final Checklist
- [x] My PR is the minimum possible work for the desired functionality
- [x] I have commented my code, particularly in hard-to-understand areas
- [ ] I have added tests that prove my fix is effective or that my feature works