@Czaki 2024-02-19
could you please add your code as benchamrk to this file:
https://github.com/napari/napari/blob/main/napari/benchmarks/benchmark_tracks_layer.py

Here is part of our documentation about benchmarking https://napari.org/stable/developers/benchmarks.html