@jules-vanaret 2024-02-19
> could you please add your code as benchamrk to this file: https://github.com/napari/napari/blob/main/napari/benchmarks/benchmark_tracks_layer.py
> 
> Here is part of our documentation about benchmarking https://napari.org/stable/developers/benchmarks.html

I'm having an issue with _asv_. I modified the current benchmark to include an additional test for updating, and ran `asv run -E existing -b TracksSuite`, which returned
```
[ 75.00%] ··· ========= =========== =========== =========== ============
              --                            n_tracks                    
              --------- ------------------------------------------------
                 size        1           10         100         1000    
              ========= =========== =========== =========== ============
                  5       14.1±6ms      n/a         n/a         n/a     
                  50      19.1±3ms    19.8±5ms      n/a         n/a     
                 500      16.6±6ms    16.6±5ms    13.0±5ms      n/a     
                 5000    30.2±10ms    25.8±9ms    25.2±9ms    20.0±4ms  
                50000     142±40ms    120±30ms    119±30ms    109±20ms  
                500000   761±400ms   582±300ms   850±300ms   1.34±0.05s 
               5000000     failed      failed      failed      failed   
              ========= =========== =========== =========== ============

[100.00%] ··· benchmark_tracks_layer.TracksSuite.time_update_layer
[100.00%] ··· ========= ============= ============= ============ ============
              --                               n_tracks                      
              --------- -----------------------------------------------------
                 size         1             10          100          1000    
              ========= ============= ============= ============ ============
                  5       4.05±0.5ms       n/a          n/a          n/a     
                  50      4.32±0.1ms   4.23±0.03ms      n/a          n/a     
                 500     5.40±0.03ms   5.12±0.03ms   4.53±0.6ms      n/a     
                 5000     15.3±0.3ms     10.5±1ms    11.9±0.3ms   11.7±0.3ms 
                50000     96.0±0.7ms    70.9±10ms    78.4±0.7ms   79.7±0.5ms 
                500000     744±10ms      584±2ms      608±8ms      630±6ms   
               5000000      failed      5.67±0.9s    7.98±0.02s     failed   
              ========= ============= ============= ============ ============
```
Unfortunately, I cannot compare to main as `asv continuous main napari_tracks_setter_optim -b TracksSuite` returns
```
· No executable found for python 3.11
· No environments selected
```
Also I don't know if it's normal, but my `.asv` folder contains only a `results` folder (so no `.asv/env` folder path exists, as specified in `asv.conf.json`).