@Czaki 2024-02-19
I have triggered benchmark on CI. Let's wait on its result.