@jules-vanaret 2024-02-19
> I have triggered benchmark on CI. Let's wait on its result.

Thank you! The benchmarks seem to be positive :)
```
| Change   | Before [0747da56]    | After [20262899]    |   Ratio | Benchmark (Parameter)                                             |
|----------|----------------------|---------------------|---------|-------------------------------------------------------------------|
| -        | 3.16±0.01ms          | 1.94±0.01ms         |    0.61 | benchmark_tracks_layer.TracksSuite.time_update_layer(500, 10)     |
| -        | 2.76±0.02ms          | 1.60±0.01ms         |    0.58 | benchmark_tracks_layer.TracksSuite.time_update_layer(50, 10)      |
| -        | 24.2±0.2ms           | 13.1±0.03ms         |    0.54 | benchmark_tracks_layer.TracksSuite.time_create_layer(5000, 100)   |
| -        | 18.2±0.4ms           | 7.72±0.07ms         |    0.42 | benchmark_tracks_layer.TracksSuite.time_create_layer(500, 100)    |
| -        | 183±0.9ms            | 74.4±0.5ms          |    0.41 | benchmark_tracks_layer.TracksSuite.time_create_layer(50000, 1000) |
| -        | 15.6±0.07ms          | 4.62±0.01ms         |    0.3  | benchmark_tracks_layer.TracksSuite.time_update_layer(5000, 100)   |
| -        | 144±0.5ms            | 35.0±0.4ms          |    0.24 | benchmark_tracks_layer.TracksSuite.time_update_layer(50000, 1000) |
| -        | 12.6±0.07ms          | 1.97±0ms            |    0.16 | benchmark_tracks_layer.TracksSuite.time_update_layer(500, 100)    |
| -        | 116±0.7ms            | 13.5±0.05ms         |    0.12 | benchmark_tracks_layer.TracksSuite.time_create_layer(5000, 1000)  |
| -        | 108±0.3ms            | 4.92±0.04ms         |    0.05 | benchmark_tracks_layer.TracksSuite.time_update_layer(5000, 1000)  |

SOME BENCHMARKS HAVE CHANGED SIGNIFICANTLY.
PERFORMANCE INCREASED.
```
The parameters range shown in the results above (`10<N_tracks<1000`, `50<N_tracks<50_000`) don't quite reflect the performance increase for large tracking datasets (i.e a `N_tracks > 100_000`) though.