@psobolewskiPhD 2024-02-19
I can confirm the bench code above is *markedly* faster on my machine locally.
With main:
```
Creating tracks with 1000000 vertices, 100000 tracks: 6.6072 seconds
Setting tracks with 1000000 vertices, 100000 tracks: 5.1887 seconds
```
with this:
```
Creating tracks with 1000000 vertices, 100000 tracks: 2.4441 seconds
Setting tracks with 1000000 vertices, 100000 tracks: 1.0332 seconds
```