@Czaki 2024-02-23
@jules-vanaret did you wait on any specific feedback, or require more explanation in the context of @brisvag request?