@brisvag 2024-02-23
> Sorry about the lack of news on my part, I had a busy week and will not be able to implement @brisvag's tests until next week (probably on monday).

Yeah, no rush, no worries :)