@jules-vanaret 2024-02-26
@brisvag I've removed the two `.copy()` and added tests to (i) check if `float64` data is handled correctly and (ii) check whether modifying a Tracks layer data inplace and calling `.refresh()` raises an error or not.
For my `float64` test, I am not sure it covers what you asked for. The test currently fails at the line `time = np.round(self._points[:, 0]).astype(np.uint)`, which does indeed cast the problematic `float64` value that I chose to the wrong value.