@brisvag 2024-02-27
Does the data on the canvas and in the input array still look the same at the end in both cases?

The reason why I suggested the `float64` test is because in the past we had issues where the input array was being modified inplace (somewhere deep in vispy) to fit the new dtype, and then converted, which unexpectedly changed your input array. If this doesn't happen, we're good on that front and we can remove the `copy()`.